requirements.txt
# Core LangChain (1.0.5+)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-google-genai>=0.0.13
langchain-experimental>=0.0.20

# Vector Store & Embeddings
faiss-cpu>=1.7.4

# Data Processing
pandas>=2.0.0
pdfplumber>=0.10.0
openpyxl>=3.1.0

# Database
sqlite-utils>=3.33.0

# Web Framework
streamlit>=1.28.0

# Environment & Config
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
